refine hide/show columns due to lazy part creation
Lazy part creation causes the content of the view is created once the
part is visible. Since an event listener is registered once the content
is created, any view doesn't receive an event before the creation.
Hence, if some columns are hidden in one view, they are not (yet) hidden
in other views that are not rendered yet.

To remedy this issue, we store the latest column hide/show in a "global"
variable.