Working draft Thread View on Linux GTK
Problems: On Linux/GTK, when displaying Thread View, EPartService always
throws an exception "application doesn not have active window".

This happens if we create a view after displaying a dialog box on
Linux/GTK. Not sure why.

Hence, the solution is to create a view first, then display a window to
choose which threads to be displayed.