Ensure compatibility with Eclipse 2019.03
We want hpcviewer to be compilable with Eclipse 2019.03 and avoid
Eclipse newer than that.
The reason is that since Eclipse 4.12 (Eclipse 2019.06), Eclipse
supports hidpi (hi definition pixels) which causes blended colors.
Since the viewer depends on pixel colors to identify the procedure, it
reports incorrect procedure names and thus incorrect statistics.

See
 https://bugs.eclipse.org/bugs/show_bug.cgi?id=565680
 https://www.eclipse.org/eclipse/news/4.12/platform.php#mac-hidpi