add org.apache.commons.text as an external plugin
We need WordUtils class to wrap long procedure name.
This class is available as an Eclipse plugin from
org.apache.commons.lang and org.apache.commons.lang3.
However, the class from these packages is deprecated, and should use
org.apache.commons.text package instead. Unfortunately, the later is not
available as an Eclipse plugin, and we need to include it manually.