Fix issue #14: extra depth
The root cause is that we mix between the depth number and the max depth
(which is the number of depths).
Assuming a tree with only one node. The depth number of this node is 0.
But the max depth is 1.
The call stack view should use the depth number, because it's what the
user sees.