Fix metric order in sparse database
In sparse database (ver 4), the order of metrics is random. Instead of
depending on the order of metric's appearance in the XML, we should
reorder metrics based on its ID.

In old XML, this doesn's matter because the order of metric appearence
is well guaranteed.

Minor update: refine id_tuple label