Working draft traces for database ver. 4
The traces can be displayed but the xml has to include TraceDB tag.
Otherwise the trace view doesn't know how to visualize it properly.

Fix: hack raw metrics to find the metric id by searching the index
one-by-one O(n) temporarily.
Assuming n (the number of metrics) is not big, it doesn't matter.