Improve the performance of the profile viewer
- Avoid iterating all the metrics to check if there's a metric value or
not within a node. Since we adopt sparse metrics (partially using
hashmap) we can just verify the size of the map (or the list for older
db).

- By default, for older db and bottom-up+flat views, we'll use
MetricValueCollectionWithStorage to store the metric values. This class
uses hashmap, thus consume less memory.

- Deprecate MetricValueCollection2. People need to use
MetricValueCollectionWithStorage instead

- Parse major and minor version at the beginning. It turns out parsing
can takes time (around 3 to 4% of the toal execution time)