Fix crash when opening flat view
The problem occurs only in release more, but doesn't happen in debug
mode inside Eclipse workspace.
This crash happens because (IMHO) we mmap too many often the file.

The solution: cache data of a certain profile (usually profile 0, which
is the summary) once. If the data is already in the cache, we don't need
to read the file again.

Minor fix:
- Use ArrayList to store the incremented nodes when building a flat tree
instead of using array to avoid allocating too much.