Fix byte buffer when reading trace.db due to insufficient buffer
Generalize the solution to read from LargeByteBuffer some bytes when
there is no enough buffer.
For instance, when a page is 1024 bytes, but we want to read 8 bytes
from index 1022, then we have to combine 2 bytes from the current page
and 6 bytes from the next page.

This commit fix for Long, Double, Float and Short