Fix id-tuple to label method
Only convert the label for rank, thread, gpu stream and gpu context.
This label can be converted later to double unit in the form of x.y to
be used in plot graph.

We don't know if we can mix between threads and gpu streams in the plot
graph. It doesn't make sense to plot in the same chart.

New:
- add a new (draft) API to get the number of variants in each kinds.