Use ConcurrentHashMap to get the cache of color & image
Since we lazily create colors and images, multiple threads can try to
access the map pair of color & image concurrently.
To avoid this, we use ConccurentHashMap instead of HashMap. And to
improve the performance, we avoid using synchronized object. Instead we
use computeIfAbsent() method to create object iff the key (the procedure
name) doesn't exist.