Draft fix issue #27 pixel-procedure mapping
If a color maps to several procedures, to avoid confusion, it's better
to concatenate multiple procedure names into a single line separated by
a comma. Otherwise, it will appear only one name in statistic view.

Also create a map between an RGB to a procedure and the color.
This is important to reuse the existing color for this procedure and
thus avoid creating another color. On Windows, the number of GDI handles
is limited.

See
 https://stackoverflow.com/questions/9723470/whats-the-upper-limit-on-gdi-objects-for-one-process-in-windows-7
 https://techcommunity.microsoft.com/t5/windows-blog-archive/pushing-the-limits-of-windows-user-and-gdi-objects-8211-part-1/ba-p/723881