Adjust the metric column width manually
The pack() method provided by SWT can be extremely slow since it
requires to check all tree items in the table.
We can bypass this method by computing the width manually. Unfortunately
SWT doesn't give us the information of the font to be used in the
header. We need to compute by guessing the font used.