Scroll to new added metric column (doesn't work on Linux/GTK)
When a user create a derived metric, the new metric column is not
revealed to the user. This gives impression that no derived metric is
created.
To solve this issue, we reveal the new column to user. Unfortunately it
doesn't work on Linux/GTK, but works fine on Mac.
Need further investigate for better portability.