Draft fix issue #34 (table is not refreshed on macOS)
Problem only occurs on macOS with Eclipse 2020.06: sorting the bottom-up
view doesn't refresh the table. This works fine on Linux/GTK.
If a user select or resize the table, the content of the table is
refreshed and the items appear.

It is unknown the root cause of the bug. It seems after the sorting the
UI thread doesn't get the items.
The solution is to delay a bit, and then expand the table to force the
table to refresh the content.