Use menu contribution to remove menu redundancy
Instead of declaring menus on the current window AND the new window
snippet, we can create menu items in the menu contribution inside
application.e4xmi file.
This doesn't allow us to create the menu folder, but at least we don't
have to declare menu items in two different places.