Windows Fix: full selection table and column width
On Windows, we have to add the attribute SWT.FULL_SELECTION to allow
users to select a row item. If this attribute is not set, the OS will
not allow to select the whole row.
This causes users unable to select a depth in call stack, and to see the
whole selection in the tables in profile view.

Other fix: adjust column width for windows by adding more space for the
percent column in Statistics view.