Partially fix self adjusted row height on mac
On Mac, the default row height can be too small and thus crop the text
in metric columns (since the font height of metric columns is bigger
than in tree column).
Forcing to expand the tree is a hack to solve this. But if the tree has
no children, this solution doesn't work.

Warning: this issue doesn't appear if we run hpcviewer within Eclipse.
If we run outside Eclipse, we'll see this bug. Sigh.