Adjust the table's row height from the start
User may set the font size to be huge, and we need to preserve this
setting. The next time hpcviewer is launched, we should adjust the row's
height accordingly.

Warning: This may cause a performance penalty since we subscribe SWT.MeasureItem event and set the height every time the event is triggered.