Fix issue #57 (null ptr exception with 2 db)
Problem: Trace view causes null pointer exception when closing one of 2
databases.
Cause: Make sure to handle "remove db" event if the trace view owns the
db.
Solution: when a "remove db" arrives, we need to check if the event is
for the trace view or for other trace view.

Impact: need to move ViewerDataEvent class from hpcviewer.ui to hpcbase so that trace viewer can access it.