Fix bug with removing a color map
The problem is when we remove a color map, we remove all the old color,
and create from scratch the map with the new color.
Since the stat table stores the old color, it barks that the color has
been disposed.

The solution is not to cache the color in the table