Hack to speed up sorting in the profile view
There is a bug in Eclipse JFace that if we refresh a table, it will try
to find the whole items in the table, including the invisible ones.

The hack is to select to the top row, and then refresh the table. In
this case, JFace only asks for the first visible region of the table