hpctree: Draft fix expanding and collapsing tree
We still use the original TreeLayer class, but since this class caches
collapsed nodes, the tree model and tree data have to also take into
account cached nodes. This makes the code looks complicated and ugly.
But it works!

Either we should create our own tree layer, or make the tree model and
tree data to be more elegant.