[major change] Fix issue #126 and use List instead of array
Issue #126 occurs when computing the depth of a flattened tree is not
consistent across nodes. This happens because some nodes are "forced" to
be in depth 1 and some are computed normally (the depth is computed from
child to the root).

When we flatten a tree, we should record the current flat level. This
level should be reset to zero when zoom-in a tree, and restore back when
we zoom-out.
The depth is then computed as the "normal" depth minus the current flat
level. This depth computation is more reliable than previous hack.