Draft fix issue #134 (dark mode on GNOME window manager)
Problem: The previous fix only works on MacOS dark mode, but not on
Linux. See issue #134

This fix is tested only on the standard dark mode of GNOME window
manager and I am not sure if it works on other themes.
The draft fix is heavily tricky using some of hundreds of knobs in
nattable to be tested which one works.
For unknown reason, this fix doesn't work on filter window only works on
the main table.
So far it works to some extend, but need to be refined further.