Draft fix issue #141 (filtering cct nodes)
The root cause is because the ScopeTreeTable stores the copy of old list
of metrics. After the filter, this the list of metrics is not the same
anymore (the content is the same, but the object references are
different).
Either we should check based on the metric id or we shouldn't cache the
list of metrics.

Right now the temp solution is to get the list of metrics from the
metric manager.