Simplify column refresh after adding a new metric
Instead of manually adding a column in the table, we should just refresh
the table AFTER the tree data is refreshed.
The advantage is that we avoid the side effect of manual column
insertion and let the nattable does everything for us.
The disadvantage is we have to manually update the position of the
sorted column.