Fix issue #138 (incorrect position of call site attribute)
Perhaps this is a nat table bug? The root cause is the image and text
decorator of the tree is managed by nat table. Somehow, if the the tree
column is so small, nat table informs that region to draw for the call
site is at the beginning of the column.

To remedy this, the decorator should be part of the tree image. So if
the tree image is not displayed (too much to the right), we shouldn't
display the attribute as well.