Draft fix issue #151 (renumbering metric index)
To support backward compatibility, it's much simpler to renumbering the
metric index inside the derived metric window instead of in the
beginning.
If we renumber in the beginning, we'll have a headache with storing the
formula as the metric index will differ from experiment.xml to the
stored one.

This may not be an issue in prof2 since we don't have metric id. But
since we need to keep backward compatibility, we should make changes
with minimal effects.