prof2: mark which line/instruction nodes to be "reduced"
Prof2 database requires to reassign contexts to move a call site as the
sibling of its parent node. Currently the reassignment causes double
reduction if the parent and grandparent nodes are line statements.

To avoid double reduction, we mark which node to be reduced, and then
perform the reassignment separately.