prof2: Add unknown and procedure call scopes, fix bottom up view
Quick fix for bottom-up view:
in prof2 a  procedure can call another non-alien procedure directly
which violates the concept of procedure frame in old prof.
When creating a caller path, we should allow this procedure as an
"enclosing call site" instead of throwing it away.
If we don't enforce this, the call chain will stop in this procedure.

It's horrible to maintain two different philosophies. It's better to
create a new bottom-up and flat view builder separately.