Fix issue #200 (incorrect cost attribution for inlines in flat view)
The root cause is because the flat id (or struct id) of inlines are the
same for different callers.
Since inlines have the same line scope flat id (because they have the
same file and line number), the viewer has to manually add the caller
(its parent) as part of the flat id.