Fix issue #187 (filter warning window)
Use a toggle check box if the user doesn't want to show the warning next
time or not.
If he/she checks the toggle, the warning won't be shown on a window
(message box) but instead on a view's message which doesn't require
people to click the annoying ok button.