Fix issue #209 (the root has no exclusive costs in the thread view)
This happens because hpcprof doesn't generate exclusive costs for the
root.
The solution is the viewer has to copy the costs from the inclusive one.
However, this causes another problem: the original hpcprof doesn't
provide information of metric partner. At the moment the viewer only
assumes that if two metrics have the same name, they must be partnered.