Fix issue #223 (incorrect cost attribution in the flat view)
This happens with prof2 database where the same procedure of the same
file and the same module has different flat id (since it has different
contexts in the CCT).
The solution is to regenerate the flat id of the line scope to depend on
the parent, not only its flat id.