fix issue #221: non-empty metrics only in graph menu
Only display non-empty metrics instead of all metrics in the context
menu for plot graphs.
The difficulty is because there is no correlation between general
metrics in metric table and metrics in metric-db. The viewer is forced
to correlate manually by matching the name.

This is not a major issue in meta.db since it has consistent metric ids
for all meta.db, profile.db and cct.db