Special case when all id-tuples are the same.
If we have:
NODE 0 THREAD 0
NODE 0 THREAD 0
...
NODE 0 THREAD 0

All the way, we shouldn't remove all the redundancies.
One solution is to keep the original id-tuple and display it as is.
Another solution is to keep just one (the last id tuple type). In this
case, just display THREAD 0.

This commit will just display as is for this corner case.