Refine flat tree for meta.db database
Two fixes:
- attribute the exclusive cost of call site instead of its line scope to
the flat counterpart.
- attribute the exclusive costs of any leaf scopes to the file and
load-module scopes.

Explanation:
1. In meta.db we need to rearrange all line scopes that have a call-site
as direct child. In this case, the call-site is "moved" as the sibling
of the line scope, but the line scope of the call site is still the
original line scope (which was the parent).
This is a mess introduced by meta.db, and I don't have a proper solution
yet. I wish meta.db generates a proper cct.

2. In old experiment.xml, all leaves are line statements.
This is not the case anymore with meta.db where any node can be a line
statement. This causes an issue when converting to flat view since all
exclusive costs for file and load-module scopes are taken from the line
statements.

Two possible solutions:
- Ensure a line statement is the leaf. This is too problematic and
against the philosophy of meta.db
- Allow flat tree visitor to attribute the exclusive costs of any leaves
to file and load-modules.

This fix adopts the 2nd solution because it's simpler and more aligned
with the meta.db philosophy.