meta.db: Fix inlined call site: mark inlined call site in top-down view
The reason there is no inline notation [i] is because the procedure
scope is taken directly from meta.db's Function section.
Unfortunately, the function table in meta.db is not exactly the same as
in experiment.xml where the function may already has the information of
inlining.

Since the procedures in this section has no information of inlining, and
only in the call site level we have this information, the top-down view
cannot display the [i] notation.

A temporary simple and portable solution is to duplicate the procedure
and flag it as inlined. This works for meta.db without chaning how we
treat experiment.xml.