Trace: Fix when cct-id of the line scope conflicts with the call site
In meta.db, sometimes we use the parent's line scope as the line of the
call site scope, but sometimes we have to create a "dummy" line for the
call site. Unfortunately the "dummy" line has the same id as the call
site and thus replace the trace id-mapping for the call site.

Solution: do not replace the call site id mapping with the "dummy" line scope