Fix callsite cost in flat tree for meta.db + fix code quality issues
meta.db requires the viewer to reassign the parent of a call site if the
parent is a line.
This happens because the viewer combines a line (a call) and its
procedure scope into one node, while in meta.db it isn't designed for
that, and the call site becomes the sibling of the line.

This causes the cost of the line of the call site to be equal to the
cost of its sibling line since they share the same context id.

Minor changes: fix code quality issues reported by sonar source