Fix meta.db plot graph by adjusting the profile position
The problem in meta.db is that the id-tuple is not sorted, and if we
sort the id-tuple in summary.db, the position of profiles in cct.db is
not sorted.
To fix this issue, temporarily we rearrange the profile in cct.db until
prof2 can sort id-tuples.

Other change: fix code issues reported by sonar source.