Draft fix issue #232 (incorrect flat tree): avoid multiple parents
Partial fix for issue #232 to avoid a node to be assigned to several
parents.
The cause of this issue is because the instruction scope is not
inherited from line scope anymore (see the previous commit).
This causes instruction scope loses one of characteristics of line
scope, which is as a call site.

Solution: If a node has a call relationship, and the parent is instruction scope, its flat id depends on the flat id of the parent.