Fix issue #233 (incorrect trace view depth)
The root cause of this is because during collecting the cpid, the viewer
collects both the call site and its line scope.
This is correct in early prof2 database but not needed in the new
meta.db file.

Solution: do not double collect trace-id of the call site. The new approach has the same id between the call (the line scope) and the callee (the procedure scope).