Refine fix issue #234 (filtering loop and/or line scopes)
Enhance the fix to refine the definition of "the same context". If two
scopes are in "the same context", we shouldn't accumulate the exclusive
cost since the parent already includes it.

However for nested loop is more complicated: the outer loop doesn't
include the cost of the inner loop. So they are basically not the same
context.