fix stability of gpu traces
when rendering traces, return the left edge of an interval that contains a time
point. previously, it returned the next point edge if the time was beyond the
midpoint of the interval.

the previous approach meant that when rendering long GPU trace intervals, there
was only a 50% chance of rendering an interval active. any time in the second
half of an active interval would be rendered as the next interval, which is
<no activity>.