Fix exception when root changes + other fixes
Additional fix issue #99 and issue #239: at the moment we avoid freeing
a root whenever we filter a database. The reason is that roots are
stored in some parts of the viewer (yeah, bad practice).
For instance, in Metric view, we store the root to get its values. If we
free the root and replace with a 'new' root, the metric view needs to be
aware.
Since there is no listener for this, we just avoid freeing a tree root.

Other fixes:
- When merging meta.db databases, careful in reading the formula '$$'.
This is allowed, not illegal.
- Use cpid instead of CCT id when storing trace id map after filtering.