Refine fix issue #99: adapt the new depth of the scope
When we propagate the trace id of the filtered node, we recompute the
depth of the "new" node (which is its ancestor) since somewhere between
the root and the new node there are some removed nodes.

Recomputing the depth is more reliable than tracking how many nodes have
been elided.