Fix issue #244 (filter and thread view issue)
The root cause is a thread part accepts an event of new filtered
database but doesn't check if this "new" filtered database is the
database in the thread part.

The problem with Eclipse events is the broadcast is global, hence
everyone listens to events that are not even part of its scope.

The solution is to check if the event is from its database by checking
the ID.