Fix issue #245 and #248 by removing unneeded tree nodes
The main problem is that meta.db requires "context reassignment" after
parsing meta.db file. This process is needed because a line statement of
a call site has the same inclusive cost as the callee function.

For example:
- line a, cost: 10
   - function b, cost: 10
      - ...

The "context reassignment" will move function b as the sibling of line a
and perform the "Reduce operation" to the line a's cost. In this example
the cost of line a becomes empty (10-10=0).

Previously, this "reduce operation" only worked with profile.db's
summary profile (i.e top-down view). But since the "original" profile.db
is not "reduced" then every time hpcviewer queries to profile.db, the
cost of line is still 10 instead of 0.

To avoid this issue, it's better to remove all empty "reduced" nodes so
that the viewer doesn't need to query its values.