Fix node depth computation, refine unit test
The depth computation is slightly incorrect because during the context
reassignment, we may remove a line scope/
For instance, originally we have:

loop a
  line x
     call f

Here, the parent of f is line x.
After the context reassignment, we merge line x and call f so that:

 loop a
  line x, call f

This commit ensures we *reset* the parent of f to loop a instead of line
x.