Draft avoiding unnecessary reopening database when filter is enabled
Reopening a database is only necessary when the database is already open
and there's a change in filter set.
In this case, we need to reopen it because some filters may be disabled
(unfilter) and we need to recompute the bottom-up and the flat view
since the metrics have changed.

For database that hasn't been opened, there is no need to reopen again,
we can skip to the filter operation directly.