Fix issue #268 (incorrect file in the call site)
The problem is because during the "calling context reassignment" we
dispose unnecessary scopes which are deleted line scopes.
During this "clean up", we also dispose the source file of the line
scope. Unfortunately, the same line scope is also used by the call site.
Hence, disposing the "deleted" line scope means disposing as well the
line of the call site.

The solution is to clone the line scope.

To do: Perhaps we should make the line scope and its file a final
constant instead?