Fix issue #267 (sorting a metric column)
The root cause is because when we sort a tree, we didn't consider the
path to the node to be sorted, but instead we just compare the depth.
This works fine with simple tree but fails with a more complicated one.

The solution is to compare the path of the two nodes.