fix the flat action in the flat view (related to issue #262)
This fix is somewhat related to issue #262 concerning flat actions.

In general we want to flatten the flat view one level (remove the parent
node) as long as it has child nodes.
For a call site is tricky since it's a leaf: if the procedure scope of
the call site has children, then it's okay to remove the call site.
However, if the procedure scope has no children (it's possible with the
new database) then we can't remove it since we may miss the exclusive
cost.