Fix issue with sorting a flattened tree
The root cause is the getPath() method is keep returning the path from
root to the node from the original tree.
When we flatten a tree, we remove the first ancestors, hence the path
has to be adapted.

The solution is to not adding the removed ancestors from the tree.
At the moment the quick fix is to make sure the nodes in the path are
not part of "flattened nodes" (its level is less than 1).

We should come up with a cleaner and faster solution. At the moment it
works just fine.