Fix flat tree, refine its unit test
The bug found only happens in the old experiment.xml database where the
procedure scope has no cost, while its children have costs.
The root cause is because the old flat tree creation code will stop
creating the flat scope counter-part if the enclosing procedure is part
of "top-down" procedure only like <program root>.

Since it doesn't want to create the flat scope counter-part, it doesn't
add the metric values to the procedure scope.

The solution is to continue to create the flat scope counter-part even
if the enclosing parent is a top-down procedure only.