Fix issue #280: incorrect sorting for a flattened tree
Issue #280: when a tree is flattened and then sorted on a different
metric column, the sorting is wrong.

The root cause is because we incorrectly collect the tree path by
including the flattened nodes in the path. the later is included because
we incorrectly compute the tree depth for flattened trees.