Use latest hpcdata to fix memory issue #68
hpcdata now will try to "dispose" the tree before filtering.

However, sometimes the table is required to be updated after the
filtering but before the table re-initialization. In this small window
gap the content of the table becomes inconsistent with the new filtered
tree and cause chaos.

To remedy this problem, at the moment we just ignore and continue until
the re-initialization completes.