Fix issue #306 (viewing traces at the edges)
We shouldn't use Min(selection.x2, width) to get the rightmost x
position. Instead, just use the selection, and let the viewer computes
the trace position.

Plus simplify the code a bit. It's just too messy for years and nobody
is eager to clean it. Need to overhaul the trace code later.