Fix issue #295 (missing metric values when merging databases)
The cause of missing values is because the way we copied metric values
is not correct for meta.db databases.
Instead of copying the content of IMetricValueCollection which may be
empty, we have to force the scope to read values from the file if it's
empty.

Need to use the latest hpcdata 7.0.0