Quick fix for hpcserver#69 issue (problem with last sample)
hpcserver proves 2 types of timestamp:
- original time stamp: the time whatever in the trace.db
- interpolate time stamp: the projection pixel-time calculated by the
server.

The first one works correctly if the trace sample includes the last
sample of the trace line. It doesn't work correctly if we don't have the
last sample.

The second one works correctly if the sample does NOT include the last
sample. It doesn't work correctly if it's the last sample.

At the moment, instead of picking only one approach, we combine both of
them:
- If we don't have the last sample, use the projected pixel-time time
stamp.
- If we have the last sample, use the original time stamp.

It's ugly, and we need a more reliable permanent solution.