ci: Upload to a single "package"
Until now the CI has uploaded multiple "packages" to the package
registry for the different OS/arch combinations that are built. This is
in part because the package registry hasn't really been used so there
was little reason to fix it. Now there is.

Uploading all the files to a single "package" (name/version) makes it
much easier to see in the GitLab UI and gives more convenient URLs to
refer to from other places.

To ensure the version is always unique, uploads are considered a kind of
deployment and are only triggered on protected tags.