Non-blocking remote database read
Do not use the standard `BufferedInputStream` which will block until it
's done reading the output from the server.
Instead, we need to use the default `ISessionRemoteSocket.read` which
will check if the stream is available or not. If it isn't available we
do read again for 5 seconds maximum.

TODO: The time needs to be configurable.