ci: Automatically package for MacOS
MacOS prefers *.dmg for distribution instead of *.zip or tarballs.
Previously this conversion had to be done by hand, this commit adds the
jobs to do the repackaging as part of the CI/CD pipeline instead.

For now, all of the produced *.dmgs are unsigned. This means they can't
be run directly without adding a security exception in MacOS. There is
much more integration necessary to make signing work properly and it
will take some time to figure out.

Co-authored-by: Laksono Adhianto <laksono@gmail.com>