Add mdformat and codespell pre-commit hooks
Mdformat is a tool for consistently formatting CommonMark files.
Codespell is a spellchecker for common mispellings in code. Both are
very useful tools for writing long fragments of prose in the repository.

Add these hooks and enable them just for the frontmatter files (README,
INSTALL, etc.). This commit also includes an autoformatting pass.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>