Add .mailmap to clean up bad Git authors
A number of Git commits in the history are attributed with useless
author information, for instance referencing the local development
system or raw username instead of a proper name and email address.

If present, Git will automatically use a .mailmap file to map authors
in the final result. This commit adds a .mailmap file with mappings for
many of the problematic authors seen in the actual Git history.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>