Add (most) missing REUSE metadata
Licensing is as usual: code is Apache-2.0, docs are CC-BY-4.0, data is
CDLA-Permissive-2.0, and random configs are CC0-1.0.
Note that Eclipse IDE files (.project, .classpath, .settings/*) are also
marked under CC0-1.0.

There are ~9 files left without annotations. 2 can't be given a license
properly, one's an assertion of the top-level license text while the
other is generated (sort of). The rest are images which need further
investigation and/or replacement.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>