Set version to 2025.1.0-SNAPSHOT
The last release was 2025.01, but Maven and Tycho have trouble with
version numbers that don't follow SemVer v1. The new version isn't
really SemVer but looks enough like it that the tools are happy. On the
plus side, all the versions across the project are now identical, which
helps to reduce version-related inconsistencies.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>