Integrate license declaration into About footer
We don't need a long text nor an extra button to just delcare that our
software is Apache 2.0. Fold it down into a single-line declaration at
the very bottom of the About dialog.

To avoid repeating ourselves (DRY principle), the footer content is
stitched together almost entirely from POM variables, the one source of
truth for the overall project license.

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>