Compact and simplify .gitignores
There are a _lot_ of .gitignore files in this repository. Clean them up
and compress them into one at the root, using generic patterns that
apply anywhere in the repository tree.

Note that proc-class.map no longer appears directly in the ignore list.
In testing this file does not appear outside the Eclipse platform
directory, which itself is under the already-ignored Maven build
directory (target/).

Signed-off-by: Jonathon Anderson <anderson.jonathonm@gmail.com>